+++ examples/commonjs/src/index.js
@@ -1,4 +1,4 @@
-const { createClient } = require("polkadot-api")
+const { createClient, CompatibilityLevel } = require("polkadot-api")
 const { ksm } = require("@polkadot-api/descriptors")
 const { start } = require("polkadot-api/smoldot")
 const { getSmProvider } = require("polkadot-api/sm-provider")
@@ -11,19 +11,22 @@ const client = createClient(getSmProvider(smoldot.addChain({ chainSpec })))
 const testApi = client.getTypedApi(ksm)
 
 async function run() {
-  const nominationsQuotaIsCompatible =
-    await testApi.apis.StakingApi.nominations_quota.isCompatible()
+  const isCompatible = (level) => level > CompatibilityLevel.Incompatible
+  const nominationsQuotaIsCompatible = isCompatible(
+    await testApi.apis.StakingApi.nominations_quota.getCompatibilityLevel(),
+  )
   console.log({ nominationsQuotaIsCompatible })
 
-  const runtime = await testApi.runtime.latest()
+  const token = await testApi.compatibilityToken
 
-  const auctionEndingIsCompatible =
-    testApi.constants.Auctions.EndingPeriod.isCompatible(runtime)
+  const auctionEndingIsCompatible = isCompatible(
+    testApi.constants.Auctions.EndingPeriod.getCompatibilityLevel(token),
+  )
 
   console.log({ auctionEndingIsCompatible })
 
   if (auctionEndingIsCompatible) {
-    console.log(testApi.constants.Auctions.EndingPeriod(runtime))
+    console.log(testApi.constants.Auctions.EndingPeriod(token))
   }
 
   await new Promise((resolve) => setTimeout(resolve, 1000))