+++ packages/cli/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Commands `add`, `remove` and `update` run codegen by default after finishing [#577](https://github.com/polkadot-api/polkadot-api/pull/577).
+- Move descriptors package out from `node_modules` [#587](https://github.com/polkadot-api/polkadot-api/pull/587).
 
 ### Removed